Moved course list into subdirectory
